requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gpiod==2.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
python-dotenv==1.1.1
pytz==2025.1
PyYAML==6.0.2
RPi.GPIO==0.7.1
suncalc==0.1.3
Werkzeug==3.1.3
